Moved ladder helper functions into helpers.py
This way we don't import the views module into things that use the
helper functions
